scraper/scripts/requirements.txt
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.18.0
firebase_admin==7.4.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.14.0
msgpack==1.1.2
packaging==26.1
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
PyJWT==2.12.1
pypdfium2==5.7.0
pytesseract==0.3.13
python-dotenv==1.2.2
requests==2.33.1
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
